requirements.txt
# Core Framework
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
email-validator==2.0.0

# Database
SQLAlchemy==2.0.20
alembic==1.12.0

# Security
bcrypt==4.0.1
PyJWT==2.8.0

# LLM and RAG Components
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-readers-docling
pdfplumber
PyPDF2
spacy
fr_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl

# UI and Visualization
matplotlib==3.8.0
Pillow==10.0.1

# Frontend Enhancements
Flask-Assets==2.0
jsmin==3.0.1
cssmin==0.2.0

# Utilities
python-dotenv==1.0.0
uuid==1.30


Werkzeug==2.3.8
